requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
comm==0.2.3
confection==0.1.5
cymem==2.0.13
datasets==4.4.2
debugpy==1.8.19
decorator==5.2.1
dill==0.4.0
executing==2.2.1
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.10.0
gensim==4.4.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mittens==0.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
preshed==3.0.12
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.3
streamlit==1.49.1
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
torch>=2.6.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
plotly==6.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
lightgbm==4.5.0
optuna==4.1.0
mlflow>=2.10.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pytest>=7.0.0
lime==0.2.0.1
shap==0.47.0
openpyxl>=3.1.0
emoji>=2.0.0
ftfy>=6.0.0
# Auth
bcrypt>=4.0.0
